requirements.txt
datasets>=2.14.0
transformers>=4.35.0
wandb>=0.16.0
aiohttp>=3.9.0
tqdm>=4.66.0
pyyaml>=6.0.0